Move grpc_server to fleet/grpc_bidi (#1864)
