+++ src/py/flwr/client/grpc_client/connection_test.py
@@ -25,7 +25,7 @@ import grpc
 
 from flwr.proto.transport_pb2 import ClientMessage, ServerMessage
 from flwr.server.client_manager import SimpleClientManager
-from flwr.server.grpc_server.grpc_server import start_grpc_server
+from flwr.server.fleet.grpc_bidi.grpc_server import start_grpc_server
 
 from .connection import grpc_connection
 
@@ -71,7 +71,7 @@ def mock_join(  # type: ignore # pylint: disable=invalid-name
 
 
 @patch(
-    "flwr.server.grpc_server.flower_service_servicer.FlowerServiceServicer.Join",
+    "flwr.server.fleet.grpc_bidi.flower_service_servicer.FlowerServiceServicer.Join",
     mock_join,
 )
 def test_integration_connection() -> None:
