+++ src/py/flwr/server/client_manager_test.py
@@ -18,7 +18,7 @@
 from unittest.mock import MagicMock
 
 from flwr.server.client_manager import SimpleClientManager
-from flwr.server.grpc_server.grpc_client_proxy import GrpcClientProxy
+from flwr.server.fleet.grpc_bidi.grpc_client_proxy import GrpcClientProxy
 
 
 def test_simple_client_manager_register() -> None:
