+++ src/py/flwr/server/fleet/grpc_bidi/grpc_client_proxy.py
@@ -21,7 +21,7 @@ from flwr import common
 from flwr.common import serde
 from flwr.proto.transport_pb2 import ClientMessage, ServerMessage
 from flwr.server.client_proxy import ClientProxy
-from flwr.server.grpc_server.grpc_bridge import GrpcBridge, InsWrapper, ResWrapper
+from flwr.server.fleet.grpc_bidi.grpc_bridge import GrpcBridge, InsWrapper, ResWrapper
 
 
 class GrpcClientProxy(ClientProxy):
