+++ src/py/flwr/server/fleet/grpc_bidi/grpc_server.py
@@ -27,8 +27,8 @@ from flwr.common.logger import log
 from flwr.proto.transport_pb2_grpc import add_FlowerServiceServicer_to_server
 from flwr.server.client_manager import ClientManager
 from flwr.server.driver.driver_servicer import DriverServicer
+from flwr.server.fleet.grpc_bidi.flower_service_servicer import FlowerServiceServicer
 from flwr.server.fleet.grpc_rere.fleet_servicer import FleetServicer
-from flwr.server.grpc_server.flower_service_servicer import FlowerServiceServicer
 
 INVALID_CERTIFICATES_ERR_MSG = """
     When setting any of root_certificate, certificate, or private_key,
