+++ src/py/flwr/server/fleet/grpc_bidi/grpc_server_test.py
@@ -23,9 +23,12 @@ from pathlib import Path
 from typing import Tuple, cast
 
 from flwr.server.client_manager import SimpleClientManager
-from flwr.server.grpc_server.grpc_server import start_grpc_server, valid_certificates
+from flwr.server.fleet.grpc_bidi.grpc_server import (
+    start_grpc_server,
+    valid_certificates,
+)
 
-root_dir = dirname(abspath(join(__file__, "../../../../..")))
+root_dir = dirname(abspath(join(__file__, "../../../../../..")))
 
 
 def load_certificates() -> Tuple[str, str, str]:
