+++ src/py/flwr/server/strategy/fedadagrad_test.py
@@ -30,7 +30,7 @@ from flwr.common import (
     parameters_to_ndarrays,
 )
 from flwr.server.client_proxy import ClientProxy
-from flwr.server.grpc_server.grpc_client_proxy import GrpcClientProxy
+from flwr.server.fleet.grpc_bidi.grpc_client_proxy import GrpcClientProxy
 
 from .fedadagrad import FedAdagrad
